Pass content-IDs of all `ContentKey` elements to `Check`s (#6859)
Introduces a new type `IdentifiedContentKey` associating a content-ID to each content-key element, and also holds the content-type.

The change became bigger than originally expected, because:
* the return type of `ContentApi.getContent(s)` had to be changed
* the `Diff` type had to be changed to provide the "from" _and_ "to" keys as `IdentifiedContentKey`s

The authz check tests had to be changed as well, because with this change, the checks need the `IdentifiedContentKey`, which is only available _after_ storage interactions. The existing test case has been split into three test cases.

New vs old storage model: the new storage model returns _all_ content IDs in `IdentifiedContentKey` - so for the content and all its namespace elements. The old storage model only returns the content ID for the content itself but not for the namespace elements.

+++ versioned/storage/store/src/main/java/org/projectnessie/versioned/storage/versionstore/BaseCommitHelper.java
@@ -591,7 +591,7 @@ class BaseCommitHelper {
                 if (op != null
                     && ex != null
                     && op.payload() == ex.payload()
-                    && contentTypeForPayload((byte) op.payload()) == NAMESPACE) {
+                    && contentTypeForPayload(op.payload()) == NAMESPACE) {
                   return ConflictResolution.ADD;
                 }
               }
